feat: add notification selectors
This commit adds selectors to the notification reducer to retrieve the following information from the state:

The currently selected filter type
All notifications
All unread notifications
These selectors make it easier to access the notification data in the Redux store and use it in the application.

The selectors are tested using the immutable library to ensure that they work correctly with Immutable.js data structures.