feat: Connect uiReducer to App component
This commit connects the uiReducer to the App component using the connect function from Redux.

The mapStateToProps function is created to map the isLoggedIn property from the uiReducer to the App component's props.

This connection allows the App component to access the isLoggedIn state from the Redux store, which can be used to conditionally render content based on the user's login statu