feat: Implement Redux store and provider in index.js
This commit implements a Redux store and provider in the index.js file.

The createStore function from Redux is used to create a store that contains the uiReducer state.

A provider is then implemented using the Provider component from React-Redux. The store is passed to the provider, which makes it available to all components within the application.

This commit is a significant step forward in the development of this project, as it allows the application to manage its state using Redux.