eat: Implement Redux state mapping in App component
This commit implements Redux state mapping in the App component using the connect function from React-Redux.

The mapStateToProps function is created to map the isLoggedIn and displayDrawer properties from the Redux state to the component's props.

This mapping allows the App component to access the isLoggedIn and displayDrawer state from the Redux store, which can be used to conditionally render content based on the user's login status and the visibility of the notification drawer.

This commit is a significant step forward in the development of this project, as it allows the App component to leverage the benefits of Redux state management.