feat: Implement Redux state mapping and action dispatching in App component
This commit implements Redux state mapping and action dispatching in the App component using the connect function from React-Redux.

The mapStateToProps function is created to map the isLoggedIn and displayDrawer properties from the Redux state to the component's props.

The mapDispatchToProps object is created to map the displayNotificationDrawer and hideNotificationDrawer action creators to the component's props.

This mapping allows the App component to access the isLoggedIn and displayDrawer state from the Redux store, and to dispatch the displayNotificationDrawer and hideNotificationDrawer actions.

This commit is a significant step forward in the development of this project, as it allows the App component to leverage the benefits of Redux state management and action dispatching.