feat: Implement Redux state mapping in Footer component
This commit implements Redux state mapping in the Footer component using the connect function from React-Redux.

The mapStateToProps function is created to map the user property from the Redux state to the component's props.

This mapping allows the Footer component to access the user state from the Redux store, which can be used to conditionally render the user's name if they are logged in.

This commit is a significant step forward in the development of this project, as it allows the Footer component to leverage the benefits of Redux state management.