feat: Implement Redux state mapping and action dispatching in Header component
This commit implements Redux state mapping and action dispatching in the Header component using the connect function from React-Redux.

The mapStateToProps function is created to map the user property from the Redux state to the component's props.

The mapDispatchToProps object is created to map the onLogout action creator to the component's props.

This mapping allows the Header component to access the user state from the Redux store, and to dispatch the onLogout action.

This commit is a significant step forward in the development of this project, as it allows the Header component to leverage the benefits of Redux state management and action dispatching.